+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.1.4] - 2025-03-21
 - Remove `--image` and `--os` flags as this is now merged into a new positional argument
 - Add `-i/--interactive` and `-t/-tty` flags for more control over the terminal
 
@@ -27,7 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - First somewhat working version
 
 <!-- next-url -->
-[Unreleased]: https://github.com/svenstaro/vmexec/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/svenstaro/vmexec/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/svenstaro/vmexec/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/svenstaro/vmexec/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/svenstaro/vmexec/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/svenstaro/vmexec/compare/v0.1.0...v0.1.1